pyproject.toml
# Note: TOML validators may show warnings for Ruff's nested configuration sections
# (e.g., [tool.ruff.lint.isort], [tool.ruff.lint.pydocstyle], etc.).
# These are false positives - the configuration is valid and works correctly with Ruff.
# The Even Better TOML extension doesn't recognize Ruff's schema, but Ruff itself validates correctly.

[project]
name = "steve"
version = "0.1.0"
description = "Claude Code agent configurations, hooks, commands, and skills"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["steve", "scripts"]

[tool.hatch.build.targets.sdist]
include = ["steve", "scripts"]

# Ruff configuration moved to ruff.toml to avoid TOML validator issues

# =============================================================================
# MyPy Configuration - Static Type Checker
# =============================================================================
[tool.mypy]
python_version = "3.10"

# Strict mode settings
strict = false
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_decorators = false
disallow_incomplete_defs = true

# Import handling
ignore_missing_imports = true
follow_imports = "silent"

# Error output
show_error_codes = true
show_column_numbers = true
pretty = true

# Exclude patterns
exclude = ["venv", ".venv", "__pycache__", "build", "dist", "node_modules"]

# Per-module overrides
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "scripts.*"
disallow_untyped_defs = false

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["scripts/tests", "steve"]
pythonpath = ["."]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
norecursedirs = ["steve/hooks", "steve/agents", "steve/commands", "steve/skills", "steve/rules", "steve/templates"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["steve", "scripts"]
branch = true
omit = ["*/__pycache__/*", "*/tests/*", "*/.venv/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]
